moved db pid logger change from settings.py to database.py
