+++ .env.example
@@ -9,6 +9,7 @@ POSTGRES_SQL_SERVICE_PORT=15432
 DATABASE_USER=postgres
 DATABASE_ADMIN=postgres
 DATABASE_PASSWORD=postgres
+LOG_DB_PID=False
 HIVE_DATABASE_NAME=hive
 HIVE_DATABASE_USER=hive
 HIVE_DATABASE_PASSWORD=hive
